Going to stop coding until this sorts out I think having 2 branches actually suck. In the end there will be 2 outcomes 1) at some point merge untested with trunk and you still get untested. After a while this will not be possible due to differences 2) don't merge and you get 2 branches. One is obviously not used
I do not do things for money or fame, but lately people started to flame me again. If you can't accept something given free and with good will then i should not try to force you to accept it. There is no oter way for me to work and as i see it, it is unaccepted even though 90% of time it has nothing to do with stability.